third_party/WebKit/LayoutTests/fast/backgrounds/repeat/margin-shorthand-expected.txt

CONSOLE WARNING: line 12: document.getMatchedCSSRules() is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4606972603138048 for more details.
This layouttest was initially there to test that shorthand property value is correct even if background-repeat property is declared before it in the style declaration. It used to test regression described in this bug. Now that access to non author stylesheet is blocked, we should instead get null when accessing the css rules on that object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iconMarginValue() is "null"
PASS successfullyParsed is true

TEST COMPLETE
